node: move local node sync cell from daemon/cmd to pkg/node/sync
This commit moves the hive cell that provides the `LocalNodeSynchronizer`
from the package `daemon/cmd` to `pkg/node/sync`.
